# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API EBAC",
        "description": "Documentation OpenAPI des API EBAC : églises, paramètres académiques et pré-inscription.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Serveur courant"
        },
        {
            "url": "https://api-ebac.severinzran.ci/api/v1",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/parametres/annees-academiques": {
            "get": {
                "tags": [
                    "Années académiques"
                ],
                "operationId": "listerAnneesAcademiques",
                "responses": {
                    "200": {
                        "description": "Liste des années académiques"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Années académiques"
                ],
                "operationId": "creerAnneeAcademique",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnneeAcademiquePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Année créée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/parametres/annees-academiques/{id}": {
            "get": {
                "tags": [
                    "Années académiques"
                ],
                "operationId": "afficherAnneeAcademique",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Année trouvée"
                    },
                    "404": {
                        "description": "Introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Années académiques"
                ],
                "operationId": "remplacerAnneeAcademique",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnneeAcademiquePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Année modifiée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Années académiques"
                ],
                "operationId": "supprimerAnneeAcademique",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Année supprimée"
                    },
                    "404": {
                        "description": "Introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Années académiques"
                ],
                "operationId": "modifierAnneeAcademique",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnneeAcademiquePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Année modifiée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AnneeAcademiquePayload": {
                "required": [
                    "libelle",
                    "date_debut",
                    "date_fin"
                ],
                "properties": {
                    "libelle": {
                        "type": "string",
                        "example": "2026-2027"
                    },
                    "date_debut": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-09-01"
                    },
                    "date_fin": {
                        "type": "string",
                        "format": "date",
                        "example": "2027-07-31"
                    },
                    "active": {
                        "type": "boolean",
                        "default": false
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Jeton retourné après la connexion et la confirmation OTP.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}